Changed test names, removed Result.IsValid, added Verifiable() method
